+++ src/Farmer/Builders/Builders.Functions.fs
@@ -391,6 +391,7 @@ type FunctionsConfig =
                         StaticWebsite = None
                         EnableHierarchicalNamespace = None
                         MinTlsVersion = None
+                        SupportsHttpsTrafficOnly = None
                         Tags = this.Tags
                         DnsZoneType = None
                         DisablePublicNetworkAccess = None